refactor(*) refactor zoomif to zoomIfOutMapExtent & interface to enum
