+++ demo/src/app/geo/feature/feature.component.ts
@@ -5,6 +5,7 @@ import {
   IgoMap,
   DataSourceService,
   LayerService,
+  OverlayAction,
   OverlayService,
   Feature,
   FeatureType,
@@ -105,6 +106,6 @@ export class AppFeatureComponent {
   }
 
   handleFeatureSelect(feature: Feature) {
-    this.overlayService.setFeatures([feature], 'zoomif');
+    this.overlayService.setFeatures([feature], OverlayAction.zoomIfOutMapExtent);
   }
 }
