+++ projects/geo/src/lib/overlay/shared/index.ts
@@ -1,3 +1,3 @@
 export * from './overlay.directive';
 export * from './overlay.service';
-export * from './overlay.interface';
+export * from './overlay.enum';
