+++ projects/geo/src/lib/overlay/shared/overlay.enum.ts
@@ -1,0 +1,6 @@
+export enum OverlayAction {
+    none,
+    move,
+    zoom,
+    zoomIfOutMapExtent
+  }
