--- projects/geo/src/lib/overlay/shared/overlay.interface.ts
@@ -1,1 +1,0 @@
-export type OverlayAction = 'none' | 'move' | 'zoom' | 'zoomif' ;
